+++ src/components/FormField/FormField.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { StyleSheet, View, TextInput, Text } from "react-native";
 import PropTypes from "prop-types";
-import { medmindBlue } from "../../utilities/styles";
+import { medmindBlue } from "../../constants/styles";
 export default class FormField extends Component {
   static propTypes = {
     header: PropTypes.string.isRequired,
