--- src/utilities/drugs.js
@@ -1,4 +1,0 @@
-// drug info such as drug types
-const drugTypes = ["Anti-Cancer", "Side Effects", "Supplements", "Other"];
-
-export { drugTypes };
