--- src/utilities/styles.js
@@ -1,4 +1,0 @@
-// reusable style props such as set colors
-const medmindBlue = "rgb(101, 192, 190)";
-
-export { medmindBlue };
